Improve day 12 speed
Reduce allocations by using a single mutable vector instead of copying the path
for every node being searched. About 3.5 times faster in release mode.

Might be even faster by using a set of visited nodes instead of the path as a
vector, but then I wouldn't be able to print the path when debugging